+++ Assets/SolarSystem/Editor/Import/SolarSystemSlice2BuildData.cs
@@ -36,6 +36,8 @@ namespace Tanvir.SolarSystem.Editor.Import
         internal Material NeptuneAtmosphereMaterial { get; set; }
         internal AirlessRockyVisualDefinition MercuryVisualDefinition { get; set; }
         internal AirlessRockyVisualDefinition MoonVisualDefinition { get; set; }
+        internal AirlessRockyVisualDefinition IoVisualDefinition { get; set; }
+        internal AirlessRockyVisualDefinition EuropaVisualDefinition { get; set; }
         internal Material OrbitMaterial { get; set; }
         internal Material SkyboxMaterial { get; set; }
         internal VolumeProfile VisualProfile { get; set; }